Add files via upload
Dang it, don't you know... window out of focus, clock (setTimeout) stops working or pauses.

Another thing to learn/implement for this is local storage. Ideally it's setup where you set your own time and endpoints eg. where to store long term (database). It's default to every minute update time stamp of active pad, and every 5 minutes, check if a new pad should be created.

It keeps getting more complex, more features. One thing too is to delete them, hey why not draggable... but it's supposed to be almost just like a text pad eg. write top to bottom.

The developing directory is the one that is not cleaned/shows thoughts... dang thanks GitHub for saving state in unsubmitted text area fields... the working directory (visible if not empty) is where the "live files" will go, the ones to use.